requirements.txt
asgiref==3.2.10
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==3.2.0
countryinfo==0.1.2
Django==3.1
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.5
django-session-timeout==0.1.0
idna==2.10
Pillow==10.0.0
psycopg2-binary==2.9.7
python-decouple==3.8
pytz==2020.1
requests==2.31.0
six==1.15.0
sqlparse==0.3.1
typing-extensions==4.7.1
urllib3==2.0.4
